Поправил названия папок(уменьшил буквы)
