+++ app/src/main/java/ru/practicum/android/diploma/data/network/JobVacancySearchApi.kt
@@ -4,8 +4,8 @@ import retrofit2.http.GET
 import retrofit2.http.Headers
 import retrofit2.http.Path
 import retrofit2.http.QueryMap
-import ru.practicum.android.diploma.data.vacancyDetail.dto.DetailResponse
-import ru.practicum.android.diploma.data.vacancyList.dto.VacanciesSearchResponse
+import ru.practicum.android.diploma.data.vacancydetail.dto.DetailResponse
+import ru.practicum.android.diploma.data.vacancylist.dto.VacanciesSearchResponse
 
 interface JobVacancySearchApi {
