--- app/src/main/java/ru/practicum/android/diploma/data/vacancyDetail/DetailRepositoryImpl.kt
@@ -1,4 +1,0 @@
-package ru.practicum.android.diploma.data.vacancyDetail
-
-class DetailRepositoryImpl {
-}
