--- app/src/main/java/ru/practicum/android/diploma/data/vacancyDetail/dto/DetailRequest.kt
@@ -1,3 +1,0 @@
-package ru.practicum.android.diploma.data.vacancyDetail.dto
-
-data class DetailRequest(val id: String)
