--- app/src/main/java/ru/practicum/android/diploma/data/vacancyDetail/dto/DetailResponse.kt
@@ -1,8 +1,0 @@
-package ru.practicum.android.diploma.data.vacancyDetail.dto
-
-import ru.practicum.android.diploma.data.Response
-import ru.practicum.android.diploma.data.vacancyDetail.dto.responseUnits.VacancyDetailDto
-
-class DetailResponse(
-    val results: VacancyDetailDto
-) : Response()
