--- app/src/main/java/ru/practicum/android/diploma/data/vacancyDetail/dto/responseUnits/Employment.kt
@@ -1,6 +1,0 @@
-package ru.practicum.android.diploma.data.vacancyDetail.dto.responseUnits
-
-data class Employment(
-    val id: String?,
-    val name: String,
-)
