--- app/src/main/java/ru/practicum/android/diploma/data/vacancyDetail/dto/responseUnits/Experience.kt
@@ -1,6 +1,0 @@
-package ru.practicum.android.diploma.data.vacancyDetail.dto.responseUnits
-
-data class Experience(
-    val id: String,
-    val name: String
-)
