--- app/src/main/java/ru/practicum/android/diploma/data/vacancyDetail/dto/responseUnits/KeySkillVacancyDetail.kt
@@ -1,5 +1,0 @@
-package ru.practicum.android.diploma.data.vacancyDetail.dto.responseUnits
-
-data class KeySkillVacancyDetail(
-    val name: String
-)
