--- app/src/main/java/ru/practicum/android/diploma/data/vacancyDetail/dto/responseUnits/Schedule.kt
@@ -1,6 +1,0 @@
-package ru.practicum.android.diploma.data.vacancyDetail.dto.responseUnits
-
-data class Schedule(
-    val id: String?,
-    val name: String
-)
