--- app/src/main/java/ru/practicum/android/diploma/data/vacancyList/SearchRepositoryImpl.kt
@@ -1,4 +1,0 @@
-package ru.practicum.android.diploma.data.vacancyList
-
-class SearchRepositoryImpl {
-}
