+++ app/src/main/java/ru/practicum/android/diploma/data/vacancydetail/DetailRepositoryImpl.kt
@@ -1,0 +1,4 @@
+package ru.practicum.android.diploma.data.vacancydetail
+
+class DetailRepositoryImpl {
+}
