+++ app/src/main/java/ru/practicum/android/diploma/data/vacancydetail/dto/DetailRequest.kt
@@ -1,0 +1,3 @@
+package ru.practicum.android.diploma.data.vacancydetail.dto
+
+data class DetailRequest(val id: String)
