+++ app/src/main/java/ru/practicum/android/diploma/data/vacancydetail/dto/DetailResponse.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.data.vacancydetail.dto
+
+import ru.practicum.android.diploma.data.Response
+import ru.practicum.android.diploma.data.vacancydetail.dto.responseunits.VacancyDetailDto
+
+class DetailResponse(
+    val results: VacancyDetailDto
+) : Response()
