+++ app/src/main/java/ru/practicum/android/diploma/data/vacancydetail/dto/responseunits/Contacts.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.data.vacancyDetail.dto.responseUnits
+package ru.practicum.android.diploma.data.vacancydetail.dto.responseunits
 
 data class Contacts(
     val email: String?,
