+++ app/src/main/java/ru/practicum/android/diploma/data/vacancydetail/dto/responseunits/Experience.kt
@@ -1,0 +1,6 @@
+package ru.practicum.android.diploma.data.vacancydetail.dto.responseunits
+
+data class Experience(
+    val id: String,
+    val name: String
+)
