+++ app/src/main/java/ru/practicum/android/diploma/data/vacancydetail/dto/responseunits/KeySkillVacancyDetail.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.data.vacancydetail.dto.responseunits
+
+data class KeySkillVacancyDetail(
+    val name: String
+)
