+++ app/src/main/java/ru/practicum/android/diploma/data/vacancydetail/dto/responseunits/Phones.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.data.vacancyDetail.dto.responseUnits
+package ru.practicum.android.diploma.data.vacancydetail.dto.responseunits
 
 data class Phones(
     val city: String,
