+++ app/src/main/java/ru/practicum/android/diploma/data/vacancylist/SearchRepositoryImpl.kt
@@ -1,0 +1,4 @@
+package ru.practicum.android.diploma.data.vacancylist
+
+class SearchRepositoryImpl {
+}
