+++ app/src/main/java/ru/practicum/android/diploma/data/vacancylist/dto/VacanciesSearchRequest.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.data.vacancyList.dto
+package ru.practicum.android.diploma.data.vacancylist.dto
 
 data class VacanciesSearchRequest(
     val queryMap: HashMap<String, String>
