+++ app/src/main/java/ru/practicum/android/diploma/data/vacancylist/dto/VacanciesSearchResponse.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.data.vacancyList.dto
+package ru.practicum.android.diploma.data.vacancylist.dto
 
 import ru.practicum.android.diploma.data.Response
 import ru.practicum.android.diploma.data.dto.responseUnits.VacancyListDto
