+++ app/src/main/java/ru/practicum/android/diploma/data/vacancylist/dto/responseunits/VacancyType.kt
